+++ spec/integration/rest/messaging/v1/brandRegistration.spec.js
@@ -65,6 +65,10 @@ describe('BrandRegistration', function() {
           'failure_reason': 'Registration error',
           'url': 'https://messaging.twilio.com/v1/a2p/BrandRegistrations/BN0044409f7e067e279523808d267e2d85',
           'brand_score': 42,
+          'identity_status': 'VERIFIED',
+          'russell_3000': true,
+          'tax_exempt_status': '501c3',
+          'skip_automatic_sec_vet': false,
           'mock': false
       };
 
@@ -105,6 +109,10 @@ describe('BrandRegistration', function() {
                   'failure_reason': 'Registration error',
                   'url': 'https://messaging.twilio.com/v1/a2p/BrandRegistrations/BN0044409f7e067e279523808d267e2d85',
                   'brand_score': 42,
+                  'identity_status': 'VERIFIED',
+                  'russell_3000': true,
+                  'tax_exempt_status': '501c3',
+                  'skip_automatic_sec_vet': false,
                   'mock': false
               }
           ]
@@ -139,6 +147,10 @@ describe('BrandRegistration', function() {
                   'failure_reason': 'Registration error',
                   'url': 'https://messaging.twilio.com/v1/a2p/BrandRegistrations/BN0044409f7e067e279523808d267e2d85',
                   'brand_score': 42,
+                  'identity_status': 'VERIFIED',
+                  'russell_3000': true,
+                  'tax_exempt_status': '501c3',
+                  'skip_automatic_sec_vet': false,
                   'mock': false
               }
           ]
@@ -178,6 +190,10 @@ describe('BrandRegistration', function() {
                   'failure_reason': 'Registration error',
                   'url': 'https://messaging.twilio.com/v1/a2p/BrandRegistrations/BN0044409f7e067e279523808d267e2d85',
                   'brand_score': 42,
+                  'identity_status': 'VERIFIED',
+                  'russell_3000': true,
+                  'tax_exempt_status': '501c3',
+                  'skip_automatic_sec_vet': false,
                   'mock': false
               }
           ]
@@ -232,6 +248,10 @@ describe('BrandRegistration', function() {
                   'failure_reason': 'Registration error',
                   'url': 'https://messaging.twilio.com/v1/a2p/BrandRegistrations/BN0044409f7e067e279523808d267e2d85',
                   'brand_score': 42,
+                  'identity_status': 'VERIFIED',
+                  'russell_3000': true,
+                  'tax_exempt_status': '501c3',
+                  'skip_automatic_sec_vet': false,
                   'mock': false
               }
           ]
@@ -292,6 +312,10 @@ describe('BrandRegistration', function() {
           'failure_reason': 'Registration error',
           'url': 'https://messaging.twilio.com/v1/a2p/BrandRegistrations/BN0044409f7e067e279523808d267e2d85',
           'brand_score': 42,
+          'identity_status': 'VERIFIED',
+          'russell_3000': true,
+          'tax_exempt_status': '501c3',
+          'skip_automatic_sec_vet': false,
           'mock': false
       };